+++ src/operations/NonUnitaryOperation.cpp
@@ -63,7 +63,7 @@ std::ostream& NonUnitaryOperation::print(
 void NonUnitaryOperation::dumpOpenQASM(std::ostream& of,
                                        const RegisterNames& qreg,
                                        const RegisterNames& creg,
-                                       uint32_t indent, bool openQASM3) const {
+                                       size_t indent, bool openQASM3) const {
   of << std::string(indent * 2, ' ');
 
   if (isWholeQubitRegister(qreg, targets.front(), targets.back()) &&
